Move helper fn's to separate file
